refactor theming
