+++ src/theme/components/TabHeading.js
@@ -1,7 +1,7 @@
 // @flow
 
 import variable from './../variables/platform';
-import { PLATFORM } from './../variables/commonColor';
+import { PLATFORM } from './../variables/common';
 
 export default (variables /* : * */ = variable) => {
   const platform = variables.platform;
